requirements.txt
openai
langchain
chromadb
reportlab
langchain-openai
langchain-chroma
langchain-text-splitters
markdown
pypdf2
python-docx
chainlit
pdfplumber
tqdm
python-dotenv